LPD-28527 Create new folder structure for tests. Move fixtures and pages where appropriate
