+++ modules/test/playwright/tests/frontend-data-set-views-web/tests/data-set-manager/fixtures/visualizationModesPageTest.ts
