+++ modules/test/playwright/tests/frontend-data-set-views-web/tests/data-set-manager/pages/DataSetsPage.ts
@@ -5,9 +5,9 @@
 
 import {Locator, Page} from '@playwright/test';
 
-import {ApiHelpers} from '../../../helpers/ApiHelpers';
-import {ApplicationsMenuPage} from '../../../pages/product-navigation-applications-menu/ApplicationsMenuPage';
-import {DEFAULT_LABEL} from '../utils/constants';
+import {ApiHelpers} from '../../../../../helpers/ApiHelpers';
+import {ApplicationsMenuPage} from '../../../../../pages/product-navigation-applications-menu/ApplicationsMenuPage';
+import {DEFAULT_LABEL} from '../../../utils/constants';
 
 export class DataSetsPage {
 	readonly apiHelpers: ApiHelpers;
