+++ modules/test/playwright/tests/frontend-data-set-views-web/tests/data-set-manager/pages/data_set/tabs/FiltersPage.ts
@@ -9,7 +9,7 @@ import {
 	IDateRangeFilter,
 	ISelectionFilterApiHeadless,
 	ISelectionFilterPicklist,
-} from '../../../utils/types';
+} from '../../../../../utils/types';
 import {DataSetPage} from '../DataSetPage';
 
 interface NewFilterModal {
